Fix "1s of future block".
setState function was using the newly set key-value pair in state to set
other variables in the same setState call. Obviously the new values
aren't going to be there - leading to false startTime and two tick()s to
update the tracked time and blocks.
Changed to now store and use startTime beforehand for the state update.